2 separated flags one for isLoading and one for hasError
