+++ app/src/main/java/com/example/georgesamuel/dubaihotels/usecases/HotelsUseCase.java
@@ -1,7 +1,7 @@
 package com.example.georgesamuel.dubaihotels.usecases;
 
 
-import com.example.georgesamuel.dubaihotels.entities.HotelsDetailsModel;
+import com.example.georgesamuel.dubaihotels.entities.HotelsResponse;
 import com.example.georgesamuel.dubaihotels.usecases.hotelsrepository.HotelsRepository;
 
 import io.reactivex.Observable;
@@ -16,7 +16,7 @@ public class HotelsUseCase {
 
     }
 
-    public Observable<HotelsDetailsModel> getHotelsDetails(){
+    public Observable<HotelsResponse> getHotelsDetails(){
        return hotelsRepository.getHotelsDetails();
     }
 }
