+++ app/src/main/java/com/example/georgesamuel/dubaihotels/usecases/hotelsrepository/HotelsRepository.java
@@ -1,11 +1,9 @@
 package com.example.georgesamuel.dubaihotels.usecases.hotelsrepository;
 
-import com.example.georgesamuel.dubaihotels.entities.HotelsDetailsModel;
+import com.example.georgesamuel.dubaihotels.entities.HotelsResponse;
 import com.example.georgesamuel.dubaihotels.presentation.features.HotelsApplication;
 import com.example.georgesamuel.dubaihotels.usecases.network.ApiClient;
 import com.example.georgesamuel.dubaihotels.usecases.network.HotelsAPI;
-
-import java.util.List;
 
 import io.reactivex.Observable;
 
@@ -17,7 +15,7 @@ public class HotelsRepository {
        hotelsAPI= ApiClient.getClient(HotelsApplication.getAppContext()).create(HotelsAPI.class);
     }
 
-    public Observable<HotelsDetailsModel> getHotelsDetails(){
+    public Observable<HotelsResponse> getHotelsDetails(){
         return hotelsAPI.getHotelsDetails();
     }
 }
