+++ app/src/main/java/com/example/georgesamuel/dubaihotels/usecases/network/HotelsAPI.java
@@ -1,6 +1,6 @@
 package com.example.georgesamuel.dubaihotels.usecases.network;
 
-import com.example.georgesamuel.dubaihotels.entities.HotelsDetailsModel;
+import com.example.georgesamuel.dubaihotels.entities.HotelsResponse;
 
 
 import retrofit2.http.GET;
@@ -9,6 +9,6 @@ import io.reactivex.Observable;
 public interface HotelsAPI {
 
     @GET("hotels")
-    Observable<HotelsDetailsModel>  getHotelsDetails ();
+    Observable<HotelsResponse>  getHotelsDetails ();
 
 }
