+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     android:id="@+id/parent_constraint_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.features.items.HotelsListActivity">
+    tools:context=".presentation.features.hotels.HotelsListActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/hotel_Items_recycler_view"
